refactor: move ResourceInitializer from ui/ to the resource layer (review #2)
ResourceInitializer is Qt-free resource bootstrap (it loads the essential LST files
FrmResolver needs) yet lived in ui/ (gecko_app), so the non-UI state/loader code
(MapLoader, DataPathLoader) had to include up into ui/ — a dependency inversion.

Move it to resource/ and into the gecko_resource library (it needs only resource/ +
format/lst). state/loader and EditorWidget now include resource/ResourceInitializer.h,
a proper downward dependency. No behavior change; gecko builds, 238 tests pass.
