+++ src/resource/ResourceInitializer.h
